moved functions from addpostservices to post services
