+++ src/services/PostServices.php
@@ -30,4 +30,20 @@ class PostServices
         $outcome .= '</section>';
          return $outcome;
     }
+
+    static public function validTitle($titleLength): bool
+    {
+        if ($titleLength > 30) {
+            return false;
+        }
+        return true;
+    }
+
+    static public function validContent($contentLength): bool
+    {
+        if ($contentLength < 50 || $contentLength > 1001) {
+            return false;
+        }
+        return true;
+    }
 }
